Remove Call/CallHolder
