+++ src/client/flow.rs
@@ -1,21 +1,25 @@
 //! A sequence of calls, such as following redirects.
 
 use std::fmt;
+use std::io::Write;
 use std::marker::PhantomData;
 
+use base64::prelude::BASE64_STANDARD;
+use base64::Engine;
 use http::header;
 use http::uri::Scheme;
 use http::{
     HeaderMap, HeaderName, HeaderValue, Method, Request, Response, StatusCode, Uri, Version,
 };
 
-use crate::body::calculate_max_input;
-use crate::ext::{HeaderIterExt, MethodExt, StatusExt};
-use crate::parser::try_parse_response;
-use crate::util::ArrayVec;
+use crate::body::{calculate_max_input, BodyReader, BodyWriter};
+use crate::ext::{HeaderIterExt, MethodExt, SchemeExt, StatusExt};
+use crate::parser::{try_parse_partial_response, try_parse_response};
+use crate::util::{log_data, ArrayVec, AuthorityExt, Writer};
 use crate::{BodyMode, Error};
 
-use super::holder::CallHolder;
+use super::amended::AmendedRequest;
+use super::MAX_RESPONSE_HEADERS;
 
 #[doc(hidden)]
 pub mod state {
@@ -55,7 +59,9 @@ pub struct Flow<B, State> {
 // pub(crate) for tests to inspect state
 #[derive(Debug)]
 pub(crate) struct Inner<B> {
-    pub call: CallHolder<B>,
+    pub request: AmendedRequest<B>,
+    pub analyzed: bool,
+    pub state: BodyState,
     pub close_reason: ArrayVec<CloseReason, 4>,
     pub should_send_body: bool,
     pub await_100_continue: bool,
@@ -71,6 +77,48 @@ impl<B> Inner<B> {
             Some(v) => v.is_redirection() && v != StatusCode::NOT_MODIFIED,
             None => false,
         }
+    }
+}
+
+#[derive(Debug, Default)]
+pub(crate) struct BodyState {
+    phase: RequestPhase,
+    writer: BodyWriter,
+    reader: Option<BodyReader>,
+    skip_method_body_check: bool,
+    allow_non_standard_methods: bool,
+    stop_on_chunk_boundary: bool,
+}
+
+impl BodyState {
+    fn need_response_body(&self) -> bool {
+        !matches!(
+            self.reader,
+            Some(BodyReader::NoBody) | Some(BodyReader::LengthDelimited(0))
+        )
+    }
+}
+
+#[derive(Clone, Copy, PartialEq, Eq)]
+enum RequestPhase {
+    SendLine,
+    SendHeaders(usize),
+    SendBody,
+}
+
+impl Default for RequestPhase {
+    fn default() -> Self {
+        Self::SendLine
+    }
+}
+
+impl RequestPhase {
+    fn is_prelude(&self) -> bool {
+        matches!(self, RequestPhase::SendLine | RequestPhase::SendHeaders(_))
+    }
+
+    fn is_body(&self) -> bool {
+        matches!(self, RequestPhase::SendBody)
     }
 }
 
@@ -131,14 +179,6 @@ impl<B, S> Flow<B, S> {
         wrapped
     }
 
-    fn call(&self) -> &CallHolder<B> {
-        &self.inner.call
-    }
-
-    fn call_mut(&mut self) -> &mut CallHolder<B> {
-        &mut self.inner.call
-    }
-
     #[cfg(test)]
     pub(crate) fn inner(&self) -> &Inner<B> {
         &self.inner
@@ -164,10 +204,21 @@ impl<B> Flow<B, Prepare> {
         let should_send_body = request.method().need_request_body();
         let await_100_continue = request.headers().iter().has_expect_100();
 
-        let call = CallHolder::new(request)?;
+        let request = AmendedRequest::new(request);
+
+        let default_body_mode = if request.method().need_request_body() {
+            BodyWriter::new_chunked()
+        } else {
+            BodyWriter::new_none()
+        };
 
         let inner = Inner {
-            call,
+            request,
+            analyzed: false,
+            state: BodyState {
+                writer: default_body_mode,
+                ..Default::default()
+            },
             close_reason,
             should_send_body,
             await_100_continue,
@@ -180,29 +231,29 @@ impl<B> Flow<B, Prepare> {
 
     /// Inspect call method
     pub fn method(&self) -> &Method {
-        self.call().request().method()
+        self.inner.request.method()
     }
 
     /// Inspect call URI
     pub fn uri(&self) -> &Uri {
-        self.call().request().uri()
+        self.inner.request.uri()
     }
 
     /// Inspect call HTTP version
     pub fn version(&self) -> Version {
-        self.call().request().version()
+        self.inner.request.version()
     }
 
     /// Inspect call headers
     pub fn headers(&self) -> &HeaderMap {
-        self.call().request().original_request_headers()
+        self.inner.request.original_request_headers()
     }
 
     /// Set whether to allow non-standard HTTP methods.
     ///
     /// By default the methods are limited by the HTTP version.
     pub fn allow_non_standard_methods(&mut self, v: bool) {
-        self.call_mut().allow_non_standard_methods(v);
+        self.inner.state.allow_non_standard_methods = v;
     }
 
     /// Add more headers to the call
@@ -213,7 +264,7 @@ impl<B> Flow<B, Prepare> {
         HeaderValue: TryFrom<V>,
         <HeaderValue as TryFrom<V>>::Error: Into<http::Error>,
     {
-        self.call_mut().request_mut().set_header(key, value)
+        self.inner.request.set_header(key, value)
     }
 
     /// Convert the call to send body despite method.
@@ -223,7 +274,7 @@ impl<B> Flow<B, Prepare> {
     /// interoperate with such services.
     pub fn send_body_despite_method(&mut self) {
         self.inner.should_send_body = true;
-        self.inner.call.convert_to_send_body();
+        self.inner.state.skip_method_body_check = true;
     }
 
     /// Continue to the next flow state.
@@ -260,35 +311,38 @@ impl<B> Flow<B, SendRequest> {
     ///
     /// The `Ok(usize)` is the number of bytes of the `output` buffer that was used.
     pub fn write(&mut self, output: &mut [u8]) -> Result<usize, Error> {
-        match &mut self.inner.call {
-            CallHolder::WithoutBody(v) => v.write(output),
-            CallHolder::WithBody(v) => v.write(&[], output).map(|r| r.1),
-            _ => unreachable!(),
-        }
+        self.maybe_analyze_request()?;
+
+        let mut w = Writer::new(output);
+        try_write_prelude(&self.inner.request, &mut self.inner.state, &mut w)?;
+
+        let output_used = w.len();
+
+        Ok(output_used)
     }
 
     /// The configured method.
     pub fn method(&self) -> &Method {
-        self.call().request().method()
+        self.inner.request.method()
     }
 
     /// The uri being requested.
     pub fn uri(&self) -> &Uri {
-        self.call().request().uri()
+        self.inner.request.uri()
     }
 
     /// Version of the request.
     ///
     /// This can only be 1.0 or 1.1.
     pub fn version(&self) -> Version {
-        self.call().request().version()
+        self.inner.request.version()
     }
 
     /// The configured headers.
     pub fn headers_map(&mut self) -> Result<HeaderMap, Error> {
-        self.call_mut().analyze_request()?;
+        self.maybe_analyze_request()?;
         let mut map = HeaderMap::new();
-        for (k, v) in self.call().request().headers() {
+        for (k, v) in self.inner.request.headers() {
             map.insert(k, v.clone());
         }
         Ok(map)
@@ -299,11 +353,7 @@ impl<B> Flow<B, SendRequest> {
     /// This is useful when the output buffer is small and we need to repeatedly
     /// call `write()` to send the entire request.
     pub fn can_proceed(&self) -> bool {
-        match &self.inner.call {
-            CallHolder::WithoutBody(v) => v.is_finished(),
-            CallHolder::WithBody(v) => v.is_body(),
-            _ => unreachable!(),
-        }
+        !self.inner.state.phase.is_prelude()
     }
 
     /// Attempt to proceed from this state to the next.
@@ -319,25 +369,160 @@ impl<B> Flow<B, SendRequest> {
             if self.inner.await_100_continue {
                 Ok(Some(SendRequestResult::Await100(Flow::wrap(self.inner))))
             } else {
-                let mut flow = Flow::wrap(self.inner);
-                flow.inner.call.analyze_request()?;
+                // TODO(martin): is this needed?
+                self.maybe_analyze_request()?;
+                let flow = Flow::wrap(self.inner);
                 Ok(Some(SendRequestResult::SendBody(flow)))
             }
         } else {
-            let call = match self.inner.call {
-                CallHolder::WithoutBody(v) => v,
-                _ => unreachable!(),
-            };
-
-            // unwrap here is ok because self.can_proceed() should check the necessary
-            // error conditions that would prevent us from converting.
-            let call_recv = call.into_receive().unwrap();
-
-            let call = CallHolder::RecvResponse(call_recv);
-            self.inner.call = call;
-
             let flow = Flow::wrap(self.inner);
             Ok(Some(SendRequestResult::RecvResponse(flow)))
+        }
+    }
+
+    pub(crate) fn maybe_analyze_request(&mut self) -> Result<(), Error> {
+        if self.inner.analyzed {
+            return Ok(());
+        }
+
+        let info = self.inner.request.analyze(
+            self.inner.state.writer,
+            self.inner.state.skip_method_body_check,
+            self.inner.state.allow_non_standard_methods,
+        )?;
+
+        if !info.req_host_header {
+            if let Some(host) = self.inner.request.uri().host() {
+                // User did not set a host header, and there is one in uri, we set that.
+                // We need an owned value to set the host header.
+
+                // This might append the port if it differs from the scheme default.
+                let value = maybe_with_port(host, self.inner.request.uri())?;
+
+                self.inner.request.set_header(header::HOST, value)?;
+            }
+        }
+
+        if let Some(auth) = self.inner.request.uri().authority() {
+            if auth.userinfo().is_some() && !info.req_auth_header {
+                let user = auth.username().unwrap_or_default();
+                let pass = auth.password().unwrap_or_default();
+                let creds = BASE64_STANDARD.encode(format!("{}:{}", user, pass));
+                let auth = format!("Basic {}", creds);
+                self.inner.request.set_header(header::AUTHORIZATION, auth)?;
+            }
+        }
+
+        if !info.req_body_header && info.body_mode.has_body() {
+            // User did not set a body header, we set one.
+            let header = info.body_mode.body_header();
+            self.inner.request.set_header(header.0, header.1)?;
+        }
+
+        self.inner.state.writer = info.body_mode;
+
+        self.inner.analyzed = true;
+        Ok(())
+    }
+}
+
+fn maybe_with_port(host: &str, uri: &Uri) -> Result<HeaderValue, Error> {
+    fn from_str(src: &str) -> Result<HeaderValue, Error> {
+        HeaderValue::from_str(src).map_err(|e| Error::BadHeader(e.to_string()))
+    }
+
+    if let Some(port) = uri.port() {
+        let scheme = uri.scheme().unwrap_or(&Scheme::HTTP);
+        if let Some(scheme_default) = scheme.default_port() {
+            if port != scheme_default {
+                // This allocates, so we only do it if we absolutely have to.
+                let host_port = format!("{}:{}", host, port);
+                return from_str(&host_port);
+            }
+        }
+    }
+
+    // Fall back on no port (without allocating).
+    from_str(host)
+}
+
+fn try_write_prelude<B>(
+    request: &AmendedRequest<B>,
+    state: &mut BodyState,
+    w: &mut Writer,
+) -> Result<(), Error> {
+    let at_start = w.len();
+
+    loop {
+        if try_write_prelude_part(request, state, w) {
+            continue;
+        }
+
+        let written = w.len() - at_start;
+
+        if written > 0 || state.phase.is_body() {
+            return Ok(());
+        } else {
+            return Err(Error::OutputOverflow);
+        }
+    }
+}
+
+fn try_write_prelude_part<Body>(
+    request: &AmendedRequest<Body>,
+    state: &mut BodyState,
+    w: &mut Writer,
+) -> bool {
+    match &mut state.phase {
+        RequestPhase::SendLine => {
+            let success = do_write_send_line(request.prelude(), w);
+            if success {
+                state.phase = RequestPhase::SendHeaders(0);
+            }
+            success
+        }
+
+        RequestPhase::SendHeaders(index) => {
+            let header_count = request.headers_len();
+            let all = request.headers();
+            let skipped = all.skip(*index);
+
+            do_write_headers(skipped, index, header_count - 1, w);
+
+            if *index == header_count {
+                state.phase = RequestPhase::SendBody;
+            }
+            false
+        }
+
+        // We're past the header.
+        _ => false,
+    }
+}
+
+fn do_write_send_line(line: (&Method, &str, Version), w: &mut Writer) -> bool {
+    w.try_write(|w| write!(w, "{} {} {:?}\r\n", line.0, line.1, line.2))
+}
+
+fn do_write_headers<'a, I>(headers: I, index: &mut usize, last_index: usize, w: &mut Writer)
+where
+    I: Iterator<Item = (&'a HeaderName, &'a HeaderValue)>,
+{
+    for h in headers {
+        let success = w.try_write(|w| {
+            write!(w, "{}: ", h.0)?;
+            w.write_all(h.1.as_bytes())?;
+            write!(w, "\r\n")?;
+            if *index == last_index {
+                write!(w, "\r\n")?;
+            }
+            Ok(())
+        });
+
+        if success {
+            *index += 1;
+        } else {
+            break;
         }
     }
 }
@@ -439,8 +624,9 @@ impl<B> Flow<B, Await100> {
         // We can always proceed out of Await100
 
         if self.inner.should_send_body {
-            let mut flow = Flow::wrap(self.inner);
-            flow.inner.call.analyze_request()?;
+            // TODO(martin): do i need this?
+            // flow.inner.call.analyze_request()?;
+            let flow = Flow::wrap(self.inner);
             Ok(Await100Result::SendBody(flow))
         } else {
             Ok(Await100Result::RecvResponse(Flow::wrap(self.inner)))
@@ -478,7 +664,22 @@ impl<B> Flow<B, SendBody> {
     /// To indicate that the body is fully sent, you call write with an `input` parameter set to `&[]`.
     /// This ends the `transfer-encoding: chunked` and ensures the state is correct to proceed.
     pub fn write(&mut self, input: &[u8], output: &mut [u8]) -> Result<(usize, usize), Error> {
-        self.inner.call.as_with_body_mut().write(input, output)
+        let mut w = Writer::new(output);
+
+        if !input.is_empty() && self.inner.state.writer.is_ended() {
+            return Err(Error::BodyContentAfterFinish);
+        }
+
+        if let Some(left) = self.inner.state.writer.left_to_send() {
+            if input.len() as u64 > left {
+                return Err(Error::BodyLargerThanContentLength);
+            }
+        }
+
+        let input_used = self.inner.state.writer.write(input, &mut w);
+        let output_used = w.len();
+
+        Ok((input_used, output_used))
     }
 
     /// Helper to avoid copying memory.
@@ -490,20 +691,25 @@ impl<B> Flow<B, SendBody> {
     /// This function is used to "report" how much of the input that has been used. It's effectively
     /// the same as the first `usize` in the pair returned by `write()`.
     pub fn consume_direct_write(&mut self, amount: usize) -> Result<(), Error> {
-        self.inner
-            .call
-            .as_with_body_mut()
-            .consume_direct_write(amount)
+        if let Some(left) = self.inner.state.writer.left_to_send() {
+            if amount as u64 > left {
+                return Err(Error::BodyLargerThanContentLength);
+            }
+        } else {
+            return Err(Error::BodyIsChunked);
+        }
+
+        self.inner.state.writer.consume_direct_write(amount);
+
+        Ok(())
     }
 
     /// Calculate the max amount of input we can transfer to fill the `output_len`.
     ///
     /// For chunked transfer, the input is less than the output.
-    pub fn calculate_max_input(&mut self, output_len: usize) -> usize {
-        let call = self.inner.call.as_with_body_mut();
-
+    pub fn calculate_max_input(&self, output_len: usize) -> usize {
         // For non-chunked, the entire output can be used.
-        if !call.is_chunked() {
+        if !self.is_chunked() {
             return output_len;
         }
 
@@ -513,9 +719,8 @@ impl<B> Flow<B, SendBody> {
     /// Test if call is chunked.
     ///
     /// This might need some processing, hence the &mut and
-    pub fn is_chunked(&mut self) -> bool {
-        let call = self.inner.call.as_with_body_mut();
-        call.is_chunked()
+    pub fn is_chunked(&self) -> bool {
+        self.inner.state.writer.is_chunked()
     }
 
     /// Check whether the request body is fully sent.
@@ -524,29 +729,17 @@ impl<B> Flow<B, SendBody> {
     /// number of bytes communicated have been sent. For chunked transfer, this becomes `true`
     /// after calling `write()` with an input of `&[]`.
     pub fn can_proceed(&self) -> bool {
-        self.inner.call.as_with_body().is_finished()
+        self.inner.state.writer.is_ended()
     }
 
     /// Proceed to the next state.
     ///
     /// Returns `None` if it's not possible to proceed. It's guaranteed that if `can_proceed()` returns
     /// `true`, this will result in `Some`.
-    pub fn proceed(mut self) -> Option<Flow<B, RecvResponse>> {
+    pub fn proceed(self) -> Option<Flow<B, RecvResponse>> {
         if !self.can_proceed() {
             return None;
         }
-
-        let call_body = match self.inner.call {
-            CallHolder::WithBody(v) => v,
-            _ => unreachable!(),
-        };
-
-        // unwrap here is ok because self.can_proceed() should check the necessary
-        // error conditions that would prevent us from converting.
-        let call_recv = call_body.into_receive().unwrap();
-
-        let call = CallHolder::RecvResponse(call_recv);
-        self.inner.call = call;
 
         Some(Flow::wrap(self.inner))
     }
@@ -572,11 +765,7 @@ impl<B> Flow<B, RecvResponse> {
         input: &[u8],
         allow_partial_redirect: bool,
     ) -> Result<(usize, Option<Response<()>>), Error> {
-        let maybe_response = self
-            .inner
-            .call
-            .as_recv_response_mut()
-            .try_response(input, allow_partial_redirect)?;
+        let maybe_response = self.do_try_response(input, allow_partial_redirect)?;
 
         let (input_used, response) = match maybe_response {
             Some(v) => v,
@@ -612,9 +801,98 @@ impl<B> Flow<B, RecvResponse> {
         Ok((input_used, Some(response)))
     }
 
+    /// Try reading response headers
+    ///
+    /// A response is only possible once the `input` holds all the HTTP response
+    /// headers. Before that this returns `None`. When the response is succesfully read,
+    /// the return value `(usize, Response<()>)` contains how many bytes were consumed
+    /// of the `input`.
+    ///
+    /// Once the response headers are succesfully read, use [`Call::into_body()`] to proceed
+    /// reading the response body.
+    fn do_try_response(
+        &mut self,
+        input: &[u8],
+        allow_partial_redirect: bool,
+    ) -> Result<Option<(usize, Response<()>)>, Error> {
+        // ~3k for 100 headers
+        let (input_used, response) = match try_parse_response::<MAX_RESPONSE_HEADERS>(input)? {
+            Some(v) => v,
+            None => {
+                // The caller decides whether to allow a partial parse.
+                if !allow_partial_redirect {
+                    return Ok(None);
+                }
+
+                // TODO(martin): I don't like this code. The mission is to be correct HTTP/1.1
+                // and this is a hack to allow for broken servers.
+                //
+                // As a special case, to handle broken servers that does a redirect without
+                // the final trailing \r\n, we try parsing the response as partial, and
+                // if it is a redirect, we can allow the request to continue.
+                let Some(mut r) = try_parse_partial_response::<MAX_RESPONSE_HEADERS>(input)? else {
+                    return Ok(None);
+                };
+
+                // A redirection must have a location header.
+                let is_complete_redirection =
+                    r.status().is_redirection() && r.headers().contains_key(header::LOCATION);
+
+                if !is_complete_redirection {
+                    return Ok(None);
+                }
+
+                // Insert a synthetic connection: close, since the connection is
+                // not valid after using a partial request.
+                debug!("Partial redirection response, insert fake connection: close");
+                r.headers_mut()
+                    .insert(header::CONNECTION, HeaderValue::from_static("close"));
+
+                (input.len(), r)
+            }
+        };
+
+        log_data(&input[..input_used]);
+
+        let http10 = response.version() == Version::HTTP_10;
+        let status = response.status().as_u16();
+
+        if status == StatusCode::CONTINUE {
+            // There should be no headers for this response.
+            if !response.headers().is_empty() {
+                return Err(Error::HeadersWith100);
+            }
+
+            return Ok(Some((input_used, response)));
+        }
+
+        let header_lookup = |name: HeaderName| {
+            if let Some(header) = response.headers().get(name) {
+                return header.to_str().ok();
+            }
+            None
+        };
+
+        let recv_body_mode =
+            BodyReader::for_response(http10, self.inner.request.method(), status, &header_lookup)?;
+
+        self.inner.state.reader = Some(recv_body_mode);
+
+        Ok(Some((input_used, response)))
+    }
+
     /// Tell if we have finished receiving the response.
     pub fn can_proceed(&self) -> bool {
-        self.inner.call.as_recv_response().is_finished()
+        self.inner.state.reader.is_some()
+    }
+
+    /// Tell if response body is closed delimited
+    ///
+    /// HTTP/1.0 does not have `content-length` to serialize many requests over the same
+    /// socket. Instead it uses socket close to determine the body is finished.
+    fn is_close_delimited(&self) -> bool {
+        let rbm = self.inner.state.reader.as_ref().unwrap();
+        matches!(rbm, BodyReader::CloseDelimited)
     }
 
     /// Proceed to the next state.
@@ -626,27 +904,17 @@ impl<B> Flow<B, RecvResponse> {
             return None;
         }
 
-        let call_body = match self.inner.call {
-            CallHolder::RecvResponse(v) => v,
-            _ => unreachable!(),
-        };
-
-        let has_response_body = call_body.need_response_body();
-        let call_body = call_body.do_into_body();
+        let has_response_body = self.inner.state.need_response_body();
 
         if has_response_body {
-            if call_body.is_close_delimited() {
+            if self.is_close_delimited() {
                 self.inner
                     .close_reason
                     .push(CloseReason::CloseDelimitedBody);
             }
 
-            self.inner.call = CallHolder::RecvBody(call_body);
-
             Some(RecvResponseResult::RecvBody(Flow::wrap(self.inner)))
         } else {
-            self.inner.call = CallHolder::RecvBody(call_body);
-
             Some(if self.inner.is_redirect() {
                 RecvResponseResult::Redirect(Flow::wrap(self.inner))
             } else {
@@ -680,7 +948,13 @@ impl<B> Flow<B, RecvBody> {
     ///
     /// The result `(usize, usize)` is `(input consumed, output buffer used)`.
     pub fn read(&mut self, input: &[u8], output: &mut [u8]) -> Result<(usize, usize), Error> {
-        self.inner.call.as_recv_body_mut().read(input, output)
+        let rbm = self.inner.state.reader.as_mut().unwrap();
+
+        if rbm.is_ended() {
+            return Ok((0, 0));
+        }
+
+        rbm.read(input, output, self.inner.state.stop_on_chunk_boundary)
     }
 
     /// Set if we are stopping on chunk boundaries.
@@ -690,10 +964,7 @@ impl<B> Flow<B, RecvBody> {
     ///
     /// Defaults to `false`
     pub fn stop_on_chunk_boundary(&mut self, enabled: bool) {
-        self.inner
-            .call
-            .as_recv_body_mut()
-            .stop_on_chunk_boundary(enabled);
+        self.inner.state.stop_on_chunk_boundary = enabled;
     }
 
     /// Tell if the reading is on a chunk boundary.
@@ -702,18 +973,33 @@ impl<B> Flow<B, RecvBody> {
     ///
     /// Only releveant if we first enabled `stop_on_chunk_boundary()`.
     pub fn is_on_chunk_boundary(&self) -> bool {
-        self.inner.call.as_recv_body().is_on_chunk_boundary()
+        let rbm = self.inner.state.reader.as_ref().unwrap();
+        rbm.is_on_chunk_boundary()
     }
 
     /// Tell which kind of mode the response body is.
     pub fn body_mode(&self) -> BodyMode {
-        self.call().body_mode()
+        self.inner.state.reader.as_ref().unwrap().body_mode()
     }
 
     /// Check if the response body has been fully received.
     pub fn can_proceed(&self) -> bool {
-        let call = self.inner.call.as_recv_body();
-        call.is_ended() || call.is_close_delimited()
+        self.is_ended() || self.is_close_delimited()
+    }
+
+    /// Tell if the response is over
+    fn is_ended(&self) -> bool {
+        let rbm = self.inner.state.reader.as_ref().unwrap();
+        rbm.is_ended()
+    }
+
+    /// Tell if response body is closed delimited
+    ///
+    /// HTTP/1.0 does not have `content-length` to serialize many requests over the same
+    /// socket. Instead it uses socket close to determine the body is finished.
+    fn is_close_delimited(&self) -> bool {
+        let rbm = self.inner.state.reader.as_ref().unwrap();
+        matches!(rbm, BodyReader::CloseDelimited)
     }
 
     /// Proceed to the next state.
@@ -777,7 +1063,7 @@ impl<B> Flow<B, Redirect> {
         };
 
         // Previous request
-        let previous = self.inner.call.request_mut();
+        let previous = &mut self.inner.request;
 
         // Unwrap is OK, because we can't be here without having read a response.
         let status = self.inner.status.unwrap();
@@ -813,7 +1099,7 @@ impl<B> Flow<B, Redirect> {
         // Next state
         let mut next = Flow::new(request)?;
 
-        let request = next.inner.call.request_mut();
+        let request = &mut next.inner.request;
 
         let keep_auth_header = match redirect_auth_headers {
             RedirectAuthHeaders::Never => false,
@@ -901,5 +1187,15 @@ impl<B> Flow<B, Cleanup> {
 impl<B, State: Named> fmt::Debug for Flow<B, State> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "Flow<{}>", State::name())
+    }
+}
+
+impl fmt::Debug for RequestPhase {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::SendLine => write!(f, "SendLine"),
+            Self::SendHeaders(_) => write!(f, "SendHeaders"),
+            Self::SendBody => write!(f, "SendBody"),
+        }
     }
 }
